コンポーネントで分ける
