+++ src/molecules/attendance-status.tsx
@@ -7,14 +7,12 @@ export type AttendanceStatusProps = {
   miss: number;
 };
 
-const AttendanceStatus = ({ onTime, late, miss }: AttendanceStatusProps) => {
-  return (
-    <div className={styles.attendanceStatus}>
-      <div className={styles.statusItem}>出席: {onTime} 回</div>
-      <div className={styles.statusItem}>遅刻: {late} 回</div>
-      <div className={styles.statusItem}>欠席: {miss} 回</div>
-    </div>
-  );
-};
+const AttendanceStatus = ({ onTime, late, miss }: AttendanceStatusProps) => (
+  <div lassName={styles.attendanceStatus}>
+    <div className={styles.statusItem}>出席: {onTime} 回</div>
+    <div className={styles.statusItem}>遅刻: {late} 回</div>
+    <div className={styles.statusItem}>欠席: {miss} 回</div>
+  </div>
+);
 
 export default AttendanceStatus;
